requirements.txt
google-generativeai
customtkinter
